backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
django-filter==23.1
psycopg2-binary==2.9.6
python-decouple==3.8
dj-database-url
Pillow==9.5.0
